# lib/api-zod/src/generated/types/createItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ThunderBill GST Invoice API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateItemRequest {
  name: string;
  hsnCode?: string | null;
  rate: number;
  gstPercent: number;
}